Add license compliance with reuse
This PR makes the project license compliance
specification with reuse lint.
https://reuse.software/

Even though it is many affected files, the pr
is scroll friendly and contains no logic.
It should be quick general overview before merging it.

Basically, add adds license headers in spdx-format,
as per the resuse-specification.
It also adds a missing reuse LICENSES file,
apache 2.0, and the .reuse/dep5 which takes
care of the licening for binary files etc.

Signed-off-by: Josef Andersson <josef.andersson@digg.se>